Move error identifier to own file to prevent side effects and fix typos
